requirements.txt
Flask==1.1.1
gunicorn==20.0.4

numpy==1.19.5
pillow==8.1.0
requests==2.22.0
Pillow==8.1.0